backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
